review ADLFile (#285)
* review ADLFile

* code rev

* adl-play example

* fix adlFile num tracks

* Fix CI

* 📝 Add docstrings to `refactor` (#286)

Docstrings generation was requested by @Raffaello.

* https://github.com/Raffaello/sdl2-hyper-sonic-drivers/pull/285#issuecomment-1873050385

The following files were modified:

* `sdl2-hyper-sonic-drivers/examples/adl-play.cpp`
* `sdl2-hyper-sonic-drivers/sdl2-hyper-sonic-drivers.cpp`
* `sdl2-hyper-sonic-drivers/src/HyperSonicDrivers/files/westwood/ADLFile.cpp`
* `sdl2-hyper-sonic-drivers/test/HyperSonicDrivers/files/westwood/TestADLFile.cpp`


* code rev

* Fix Build

* Fix SonarCloud

* Fix ADL v3 header type word instead of byte

* update Version patch

* codeRabbit rev

* Fix Audio Driver Play track type to uint16_t

* code rev

* minor improve of cross-OS code compatibility

---------

+++ sdl2-hyper-sonic-drivers/test/HyperSonicDrivers/drivers/TestIAudioDriver.cpp
@@ -8,7 +8,7 @@ namespace HyperSonicDrivers::drivers
     {
     public:
         IAudioDriverMock(const std::shared_ptr<devices::IDevice>& device) : IAudioDriver(device) {}
-        void play(const uint8_t track) noexcept override {};
+        void play(const uint16_t track) noexcept override {};
         void stop() noexcept override {};
         bool isPlaying() const noexcept override { return false; };
     };
